==> 1 build_base_dataset.ipynb <==
## Cargar Dependencias

import os
import glob
import json
import seaborn as sns

api_data_path = "./api-data/"

## Leer todos los json de los pacientes

patient_files = glob.glob(os.path.join(api_data_path, "*.json"))
patient_data = [
    json.loads(open(file, "r").read()) for file in patient_files
]

## Definir funciones

def hypertension_medicine(medicamentos):
    """
    This is a function that standarizes the drugs used in the database and returns if patient is 
    under hypertensive treatment or not
    """
    nombres_medicamentos = { #{'medicamento':'', 'dosis': , 'unidades': 'mg'}
        ' (metoprolol succinato) 50 mg' : [{'medicamento':'metoprolol', 'dosis': 50, 'unidades': 'mg'}],
        'acetazolamida (acetazolamida) 250 mg' : [{'medicamento':'acetazolamida', 'dosis': 250, 'unidades': 'mg'} ],
        'adalat oros (nifedipino) 20 mg' : [{'medicamento':'nifedipino', 'dosis': 20, 'unidades': 'mg'}],
        'adalat oros (nifedipino) 30 mg' : [{'medicamento':'nifedipino', 'dosis': 30, 'unidades': 'mg'}],
        'adalat oros (nifedipino) 60 mg' : [{'medicamento':'nifedipino', 'dosis': 60, 'unidades': 'mg'}],
        'aldactone  (espironolactona) 100mg' : [{'medicamento':'espironolactona', 'dosis': 100, 'unidades': 'mg'}],
        'aldactone  (espironolactona) 25mg' : [{'medicamento':'espironolactona', 'dosis': 25, 'unidades': 'mg'}],
        'alfametildopa (alfametildopa) 250 mg' : [{'medicamento':'alfametildopa', 'dosis': 250, 'unidades': 'mg'}],
        'amdipin h (amlodipino+hidroclorotiazida) 5 mg/12,5mg' : [{'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'},{'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'amlodipino (amlodipino) 10 mg' : [{'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}],
        'amlodipino (amlodipino) 5 mg' : [{'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'}],
        'aprovasc (irbersartan/amlodipino) 150/10 mg' : [{'medicamento':'irbesartan', 'dosis': 150, 'unidades': 'mg'},{'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}],
        'aprovasc (irbesartan 150 + amlodipino 5 mg) 150 mg' : [{'medicamento':'irbesartan', 'dosis': 150, 'unidades': 'mg'},{'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'}],
        'aprovasc (irbesartan 300 + amlodipino 10 mg) 300mg' : [{'medicamento':'irbesartan', 'dosis': 300, 'unidades': 'mg'},{'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}],
        'aprovasc (irbesartan/amlodipino) 300/5 mg' : [{'medicamento':'irbesartan', 'dosis': 300, 'unidades': 'mg'},{'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'}],
        'aprovel (irbesartan) 150mg' : [{'medicamento':'irbesartan', 'dosis': 150, 'unidades': 'mg'}],
        'aprovel (irbesartan) 300mg' : [{'medicamento':'irbesartan', 'dosis': 300, 'unidades': 'mg'}],
        'atacand  (candesartan) 8mg' : [{'medicamento':'candesartan', 'dosis': 8, 'unidades': 'mg'}],
        'atacand (candesartan) 16mg' : [{'medicamento':'candesartan', 'dosis': 16, 'unidades': 'mg'}],
        'atacand plus (candesartan+hidroclorotiazida) 16mg/12,5mg' : [{'medicamento':'candesartan', 'dosis': 16, 'unidades': 'mg'},{'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'atenolol (atenolol) 100 mg' : [{'medicamento':'atenolol', 'dosis': 100, 'unidades': 'mg'}],
        'atpure (atenolol) 25 mg' : [{'medicamento':'atenolol', 'dosis': 25, 'unidades': 'mg'}],
        'atpure (atenolol) 50 mg' : [{'medicamento':'atenolol', 'dosis': 50, 'unidades': 'mg'}],
        'beloc (metoprolol tartrato) 50 mg' : [{'medicamento':'metoprolol', 'dosis': 50, 'unidades': 'mg'}],
        'betaloc (metoprolol tartrato) 100 mg' : [{'medicamento':'metoprolol', 'dosis': 100, 'unidades': 'mg'}],
        'betaloc (metoprolol tartrato) 50 mg' : [{'medicamento':'metoprolol', 'dosis': 50, 'unidades': 'mg'}],
        'betaloc zoc (metoprolol succinato) 100 mg' : [{'medicamento':'metoprolol', 'dosis': 100, 'unidades': 'mg'}],
        'betaloc zoc (metoprolol succinato) 200 mg' : [{'medicamento':'metoprolol', 'dosis': 200, 'unidades': 'mg'}],
        'betaloc zoc (metoprolol succinato) 25 mg' : [{'medicamento':'metoprolol', 'dosis': 25, 'unidades': 'mg'}],
        'betaloc zoc (metoprolol succinato) 50 mg' : [{'medicamento':'metoprolol', 'dosis': 50, 'unidades': 'mg'}],
        'betoprolol (metoprolol tartrato) 100 mg' : [{'medicamento':'metoprolol', 'dosis': 100, 'unidades': 'mg'}],
        'bifril (zofenopril) 30 mg' : [{'medicamento':'zofenopril', 'dosis': 30, 'unidades': 'mg'}],
        'bipreterax<br> (perindopril/indapamida) 5/1,25' : [{'medicamento':'perindopril', 'dosis': 5, 'unidades': 'mg'},{'medicamento':'indapamida', 'dosis': 1.25, 'unidades': 'mg'}],
        'candam h (candesartan/amlodipino/hidroclorotiazida) 32/5/12.5 mg' : [{'medicamento':'candesartan', 'dosis': 32, 'unidades': 'mg'},{'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'},{'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'candesartan (candesartan) 16mg' : [{'medicamento':'candesartan', 'dosis': 16, 'unidades': 'mg'}],
        'candesartan (candesartan) 32 mg' : [{'medicamento':'candesartan', 'dosis': 32, 'unidades': 'mg'}],
        'candesartan (candesartan) 8mg' : [{'medicamento':'candesartan', 'dosis': 8, 'unidades': 'mg'}],
        'capoten (captopril ) 25mg' : [{'medicamento':'captopril', 'dosis': 25, 'unidades': 'mg'}],
        'capoten (captopril ) 50mg': [{'medicamento':'captopril', 'dosis': 50, 'unidades': 'mg'}],
        'captopril  (captopril ) 25mg' : [{'medicamento':'captopril', 'dosis': 25, 'unidades': 'mg'}],
        'captopril  (captopril ) 50mg' : [{'medicamento':'captopril', 'dosis': 50, 'unidades': 'mg'}],
        'cardik a (valsartan + amlodipino) 160/5' : [{'medicamento':'valsartan', 'dosis': 160, 'unidades': 'mg'},{'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'}],
        'cardik in (valsartán/indapamida) 160/1.25' : [{'medicamento':'valsartan', 'dosis': 160, 'unidades': 'mg'},{'medicamento':'indapamida', 'dosis': 1.25, 'unidades': 'mg'}],
        'cardik3 (valsartan/amlodipino/hctz) 160/10/12.5' : [{'medicamento':'valsartan', 'dosis': 160, 'unidades': 'mg'},{'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'},{'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'cardik3 (valsartan/amlodipino/hctz) 160/5/12.5' : [{'medicamento':'valsartan', 'dosis': 160, 'unidades': 'mg'},{'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'},{'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'cardiocap (telmisartán) 80' : [{'medicamento':'telmisartan', 'dosis': 80 , 'unidades': 'mg'}],
        'cardiocap h (telmisartán+hctz) 80/12.5' : [{'medicamento':'telmisartan', 'dosis': 80 , 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'carvedil (carvedilol) 12.5 mg' : [{'medicamento':'carvedilol', 'dosis': 12.5, 'unidades': 'mg'}],
        'carvedilol (carvedilol) 12.5 mg' : [{'medicamento':'carvedilol', 'dosis': 12.5, 'unidades': 'mg'}],
        'carvedilol (carvedilol) 25 mg' : [{'medicamento':'carvedilol', 'dosis': 25, 'unidades': 'mg'}],
        'carvedilol (carvedilol) 6.25 mg' : [{'medicamento':'carvedilol', 'dosis': 6.25, 'unidades': 'mg'}],
        'catapresan (clonidina) 150mcg'  : [{'medicamento':'clonidina', 'dosis': 150, 'unidades': 'mcg'}],
        'clonidina (clonidina) 150mcg'  : [{'medicamento':'clonidina', 'dosis': 150, 'unidades': 'mcg'}],
        'clortalidona (clortalidona) 25' : [{'medicamento':'clortalidona', 'dosis': 25, 'unidades': 'mg'}],
        'clortalidona (clortalidona) 50' : [{'medicamento':'clortalidona', 'dosis': 50, 'unidades': 'mg'}],
        'coaprovel (irbesartan+ hidroclotiazida) 150mg/12,5mg' : [{'medicamento':'irbesartan', 'dosis': 150 , 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'coaprovel (irbesartan+ hidroclotiazida) 300mg/12,5mg' : [{'medicamento':'irbesartan', 'dosis': 300 , 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'coaprovel (irbesartan+ hidroclotiazida) 300mg/25mg' : [{'medicamento':'irbesartan', 'dosis': 300 , 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 25, 'unidades': 'mg'}],
        'coaprovel (irbesartan+ hidroclotiazida) 300mg/50mg' : [{'medicamento':'irbesartan', 'dosis': 300 , 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 50, 'unidades': 'mg'}],
        'concor (bisoprolol fumarato) 1.25 mg' : [{'medicamento':'bisoprolol', 'dosis': 1.25, 'unidades': 'mg'}],
        'concor (bisoprolol fumarato) 10 mg' : [{'medicamento':'bisoprolol', 'dosis': 10, 'unidades': 'mg'}],
        'concor (bisoprolol fumarato) 2.5 mg' : [{'medicamento':'bisoprolol', 'dosis': 2.5, 'unidades': 'mg'}],
        'concor (bisoprolol fumarato) 5 mg' : [{'medicamento':'bisoprolol', 'dosis': 5, 'unidades': 'mg'}],
        'concor am (bisoprolol/amlodiìno) 5/5 mg' : [{'medicamento':'bisoprolol', 'dosis': 5 , 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'}],
        'coryol (carvedilol) 12.5 mg' : [{'medicamento':'carvedilol', 'dosis': 12.5, 'unidades': 'mg'}],
        'coryol (carvedilol) 6.25 mg' : [{'medicamento':'carvedilol', 'dosis': 6.25, 'unidades': 'mg'}],
        'coveram 5/5 (perindopril/amlodipino) 5/5 mg' : [{'medicamento':'perindopril', 'dosis': 5 , 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'}],
        'coveratrix (coveratrix) 10/2.5/10' : [{'medicamento':'perindopril', 'dosis': 10 , 'unidades': 'mg'}, {'medicamento':'indapamida', 'dosis': 2.5, 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}],
        'coveratrix (perindopril/amlo) 10/2.5/10' : [{'medicamento':'perindopril', 'dosis': 10 , 'unidades': 'mg'}, {'medicamento':'indapamida', 'dosis': 2.5, 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}],
        'coveratrix (perindopril/amlo) 5/2.5/10' : [{'medicamento':'perindopril', 'dosis': 5 , 'unidades': 'mg'}, {'medicamento':'indapamida', 'dosis': 2.5, 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}],
        'cozaar (losartan) 100mg' : [{'medicamento':'losartan', 'dosis': 100, 'unidades': 'mg'}],
        'cozaar (losartan) 50mg' : [{'medicamento':'carvedilol', 'dosis': 50, 'unidades': 'mg'}],
        'dilatrend (carvedilol) 12.5 mg' : [{'medicamento':'carvedilol', 'dosis': 12.5, 'unidades': 'mg'}],
        'dilatrend (carvedilol) 25 mg' : [{'medicamento':'carvedilol', 'dosis': 25, 'unidades': 'mg'}],
        'dilatrend (carvedilol) 6.25 mg' : [{'medicamento':'carvedilol', 'dosis': 6.25, 'unidades': 'mg'}],
        'diltiazem (diltiazem clorhidrato) 120mg' : [{'medicamento':'diltiazem', 'dosis': 120, 'unidades': 'mg'}],
        'diltiazem (diltiazem clorhidrato) 180mg' : [{'medicamento':'diltiazem', 'dosis': 180, 'unidades': 'mg'}],
        'diltiazem (diltiazem clorhidrato) 60mg' : [{'medicamento':'diltiazem', 'dosis': 60, 'unidades': 'mg'}],
        'diovan  (valsartan+hidroclorotiazida) 160mg/12,5' : [{'medicamento':'valsartan', 'dosis': 160 , 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'diovan  (valsartan+hidroclorotiazida) 80mg/12,5mg' : [{'medicamento':'valsartan', 'dosis': 80 , 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'diovan (valsartan ) 160mg' : [{'medicamento':'valsartan', 'dosis': 160 , 'unidades': 'mg'}],
        'diovan (valsartan) 320' : [{'medicamento':'valsartan', 'dosis': 320 , 'unidades': 'mg'}],
        'diovan (valsartan) 80 mg' : [{'medicamento':'valsartan', 'dosis': 80 , 'unidades': 'mg'}],
        'diovan (valsartan) 80mg' : [{'medicamento':'valsartan', 'dosis': 80 , 'unidades': 'mg'}],
        'diovan (valsartan) tableta 320 mg' : [{'medicamento':'valsartan', 'dosis': 320 , 'unidades': 'mg'}],
        'diovan hct (valsartán+hidroclorotiazida) 160/25' : [{'medicamento':'valsartan', 'dosis': 160 , 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'diurex (indapamida) 2.5mg' : [{'medicamento':'indapamida', 'dosis': 2.5 , 'unidades': 'mg'}],
        'enalapril (enalapril) 10mg' : [{'medicamento':'enalapril', 'dosis': 10 , 'unidades': 'mg'}],
        'enalapril (enalapril) 20mg' : [{'medicamento':'enalapril', 'dosis': 20 , 'unidades': 'mg'}],
        'enalapril (enalapril) 5mg' : [{'medicamento':'enalapril', 'dosis': 5 , 'unidades': 'mg'}],
        'eplenorona (eplenorona) 50' : [{'medicamento':'eplenorona', 'dosis': 50 , 'unidades': 'mg'}],
        'eprosartán (eprosartan) '  : [{'medicamento':'eprosartan', 'dosis': 0 , 'unidades': 'mg'}],
        'espironolactona (espironolactona) 25mg'  : [{'medicamento':'espironolactona', 'dosis': 25, 'unidades': 'mg'}],
        'eucoran (amlodipino) 5 mg' : [{'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'}],
        'exforge (160/10) 160/10 mg' : [{'medicamento':'valsartan', 'dosis': 160 , 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}],
        'exforge (valsartan/amlodipino) 160/10' : [{'medicamento':'valsartan', 'dosis': 160 , 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}],
        'exforge (valsartan/amlodipino) 160/5' : [{'medicamento':'valsartan', 'dosis': 160 , 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'}],
        'exforge (valsartan/amlodipino) 320/10' : [{'medicamento':'valsartan', 'dosis': 320 , 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}],
        'exforge 320/10 (valsartan/amlodipino) 320/10' : [{'medicamento':'valsartan', 'dosis': 320 , 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}],
        'exforge 80/5 (valsartan/amlodipino) 80/5' : [{'medicamento':'valsartan', 'dosis': 80 , 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'}],
        'exforge hct (amlodipino/valsartan/hidroclorotriazida) 5mg/160mg/12.5mg' : [{'medicamento':'valsartan', 'dosis': 160 , 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'exforge hct (valsartan/amlo/hctz) 160/10/12.5' : [{'medicamento':'valsartan', 'dosis': 160 , 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'exforge hct (valsartan/amlodipino/hctz) 160/5/12.5' : [{'medicamento':'valsartan', 'dosis': 160 , 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'exforge hctz (antihipertensivo) 320/10/25' : [{'medicamento':'valsartan', 'dosis': 320 , 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 25, 'unidades': 'mg'}],
        'exforge hctz (valsartan/amlodipino/hidroclorotiazida) 320/10/25' : [{'medicamento':'valsartan', 'dosis': 320 , 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 25, 'unidades': 'mg'}],
        'furosemida (furosemida) 40mg' : [{'medicamento':'furosemida', 'dosis': 40, 'unidades': 'mg'}],
        'gotland neviten (nebivolol clorhidrato) 5 mg' : [{'medicamento':'nebivolol', 'dosis': 5, 'unidades': 'mg'}],
        'hidroclorotiazida\xa0 (hidroclorotiazida\xa0) 25mg' : [{'medicamento':'hidroclorotiazida', 'dosis': 25, 'unidades': 'mg'}],
        'hidroclorotiazida\xa0 (hidroclorotiazida\xa0) 50mg' : [{'medicamento':'hidroclorotiazida', 'dosis': 50, 'unidades': 'mg'}],
        'hidroten (clortalidona) ' : [{'medicamento':'clortalidona', 'dosis': 0, 'unidades': 'mg'}],
        'hidroten (clortalidona) 25' : [{'medicamento':'clortalidona', 'dosis': 25, 'unidades': 'mg'}],
        'hyzaar (losartan+hidroclorotiazida) 50/12.5' : [{'medicamento':'losartan', 'dosis': 50 , 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'hyzaar (losartan+hidroclorotiazida) 50mg/12,5mg' : [{'medicamento':'losartan', 'dosis': 50 , 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'hyzaar&nbsp; (losartan+hidroclorotiazida) 100/12.5' : [{'medicamento':'losartan', 'dosis': 100 , 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'iltux 40 (olmesartán) 40 mg' : [{'medicamento':'olmesartan', 'dosis': 40, 'unidades': 'mg'}],
        'iltux am (iltux am) 40/5' : [{'medicamento':'olmesartan', 'dosis': 40, 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'}],
        'inderal (propanolol clorhidrato) 40 mg' : [{'medicamento':'propanolol', 'dosis': 40, 'unidades': 'mg'}],
        'inderal (propranolol clorhidrato) 40 mg' : [{'medicamento':'propanolol', 'dosis': 40, 'unidades': 'mg'}],
        'indpamida (indapamida) 1.5mg' : [{'medicamento':'indapamida', 'dosis': 1.5, 'unidades': 'mg'}],
        'irbersartan+amlodipino (irbersartan+amlodipino) 300/5 mg' : [{'medicamento':'irbesartan', 'dosis': 300, 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'}],
        'irbesartan (irbesartan) 150mg' : [{'medicamento':'irbesartan', 'dosis': 150, 'unidades': 'mg'}],
        'irbesartan (irbesartan) 300mg' : [{'medicamento':'irbesartan', 'dosis': 300, 'unidades': 'mg'}],
        'irbesartan (irbesartan) 75mg' : [{'medicamento':'irbesartan', 'dosis': 75, 'unidades': 'mg'}],
        'irbesartan/hidroclorotiazida (irbesartan/hidroclorotiazida) 300mg/25mg' : [{'medicamento':'irbesartan', 'dosis': 300, 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 25, 'unidades': 'mg'}],
        'isoptin (verapamilo) 80mg ' : [{'medicamento':'verapamilo', 'dosis': 80, 'unidades': 'mg'}],
        'isoptin sr (verapamilo) 120mg' : [{'medicamento':'verapamilo', 'dosis': 120, 'unidades': 'mg'}],
        'isoptin sr (verapamilo) 240mg' : [{'medicamento':'verapamilo', 'dosis': 240, 'unidades': 'mg'}],
        'lasix (furosemida) 40mg' : [{'medicamento':'furosemida', 'dosis': 40, 'unidades': 'mg'}],
        'lisinopril  (lisinopril ) 10mg' : [{'medicamento':'lisinopril', 'dosis': 10, 'unidades': 'mg'}],
        'lisinopril  (lisinopril ) 20mg' : [{'medicamento':'lisinopril', 'dosis': 20, 'unidades': 'mg'}],
        'lisinopril  (lisinopril ) 5mg' : [{'medicamento':'lisinopril', 'dosis': 5, 'unidades': 'mg'}],
        'lopresor (metoprolol tartrato) 100 mg' : [{'medicamento':'metoprolol', 'dosis': 100, 'unidades': 'mg'}],
        'lopresor (metoprolol tartrato) 50 mg' : [{'medicamento':'metoprolol', 'dosis': 50, 'unidades': 'mg'}],
        'losartan (losartan) 100 mg' : [{'medicamento':'losartan', 'dosis': 100, 'unidades': 'mg'}],
        'losartan (losartan) 50mg': [{'medicamento':'losartan', 'dosis': 50, 'unidades': 'mg'}],
        'losartan+hidroclorotiazida (losartan+hidroclorotiazida) 100/25' : [{'medicamento':'losartan', 'dosis': 100 , 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 25, 'unidades': 'mg'}],
        'losartan+hidroclorotiazida (losartan+hidroclorotiazida) 100mg/12,5mg' : [{'medicamento':'losartan', 'dosis': 100 , 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'losartan+hidroclorotiazida (losartan+hidroclorotiazida) 100mg/25mg' : [{'medicamento':'losartan', 'dosis': 100 , 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 25, 'unidades': 'mg'}],
        'losartan+hidroclorotiazida (losartan+hidroclorotiazida) 50/12.5' : [{'medicamento':'losartan', 'dosis': 50 , 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'losartan+hidroclorotiazida (losartan+hidroclorotiazida) 50mg/12,5mg' : [{'medicamento':'losartan', 'dosis': 50 , 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'losartan/hidroclorotiazida (losartan/hidroclorotiazida) 100/25' : [{'medicamento':'losartan', 'dosis': 100 , 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 25, 'unidades': 'mg'}],
        'losartan/hidroclorotiazida (losartan/hidroclorotiazida) 50/12.5' : [{'medicamento':'losartan', 'dosis': 50 , 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'metoprolol (metoprolol tartrato) 100 mg' : [{'medicamento':'metoprolol', 'dosis': 100, 'unidades': 'mg'}],
        'metoprolol (metoprolol tartrato) 25 mg' : [{'medicamento':'metoprolol', 'dosis': 25, 'unidades': 'mg'}],
        'metoprolol (metoprolol tartrato) 50 mg' : [{'medicamento':'metoprolol', 'dosis': 50, 'unidades': 'mg'}],
        'metpure xl (metoprolol succinato) 25 mg' : [{'medicamento':'metoprolol', 'dosis': 25, 'unidades': 'mg'}],
        'metpure xl (metoprolol succinato) 50 mg' : [{'medicamento':'metoprolol', 'dosis': 50, 'unidades': 'mg'}],
        'micardis (telmisartan) 40mg' : [{'medicamento':'telmisartan', 'dosis': 40, 'unidades': 'mg'}],
        'micardis (telmisartan) 80mg' : [{'medicamento':'telmisartan', 'dosis': 80, 'unidades': 'mg'}],
        'micardis amlo (telmisartan+amlodipino) 80/10' : [{'medicamento':'telmisartan', 'dosis': 80, 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}],
        'micardis amlo (telmisartan+amlodipino) 80/5' : [{'medicamento':'telmisartan', 'dosis': 80, 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'}],
        'micardis plus (telmisartan+hidroclorotiazida) 80mg/12,5mg' : [{'medicamento':'telmisartan', 'dosis': 80, 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'micardis plus (telmisartan+hidroclorotiazida) 80mg/25mg' : [{'medicamento':'telmisartan', 'dosis': 80, 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 25, 'unidades': 'mg'}],
        'minipres (prazosin) 1mg' : [{'medicamento':'prazosin', 'dosis': 1, 'unidades': 'mg'}],
        'minoxiten (minoxidil) 10 mg' : [{'medicamento':'minoxidil', 'dosis': 10, 'unidades': 'mg'}],
        'moduretic (amilorida/hidrolorotiazida) 5/50' : [{'medicamento':'amilorida', 'dosis': 5, 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 50, 'unidades': 'mg'}],
        'monopril (fosinopril) 20mg' : [{'medicamento':'fosinopril', 'dosis': 20, 'unidades': 'mg'}],
        'nabila (nebivolol clorhidrato) 10 mg' : [{'medicamento':'nebivolol', 'dosis': 10, 'unidades': 'mg'}],
        'nabila (nebivolol clorhidrato) 2.5' : [{'medicamento':'nebivolol', 'dosis': 2.5, 'unidades': 'mg'}],
        'nabila (nebivolol clorhidrato) 5 mg' : [{'medicamento':'nebivolol', 'dosis': 5, 'unidades': 'mg'}],
        'natrilix (indapamida) 1.5mg' : [{'medicamento':'indapamida', 'dosis': 1.5, 'unidades': 'mg'}],
        'nebilet (nebivolol clorhidrato) 5 mg' : [{'medicamento':'nebivolol', 'dosis': 5, 'unidades': 'mg'}],
        'nebilet (nebivolol) 10 mg' : [{'medicamento':'nebivolol', 'dosis': 10, 'unidades': 'mg'}],
        'nebivolol (nebivolol clorhidrato) 5 mg' : [{'medicamento':'nebivolol', 'dosis': 5, 'unidades': 'mg'}],
        'nifedipino (nifedipino) 30 mg' : [{'medicamento':'nifedipino', 'dosis': 30, 'unidades': 'mg'}],
        'nifedipino (nifedipino) 60 mg' : [{'medicamento':'nifedipino', 'dosis': 60, 'unidades': 'mg'}],
        'nimodipino (nimodipino) 30 mg' : [{'medicamento':'nimodipino', 'dosis': 30, 'unidades': 'mg'}],
        'norvas (amlodipino) 10 mg' : [{'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}],
        'norvas (amlodipino) 5 mg' : [{'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'}],
        'olmesartan (olmesartán medoxomil) 20mg' : [{'medicamento':'olmesartan', 'dosis': 20, 'unidades': 'mg'}],
        'olmetec (olmesartán medoxomil) 40mg' : [{'medicamento':'olmesartan', 'dosis': 40, 'unidades': 'mg'}],
        'olmetec hct (olmesartán/hidroclorotiazida) 20mg/12,5mg' : [{'medicamento':'olmesartan', 'dosis': 20, 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'olmetec hct (olmesartán/hidroclorotiazida) 40mg/12,5mg' : [{'medicamento':'olmesartan', 'dosis': 40, 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'olmetec/amlo (olmesartan/alodipino) 40/5' : [{'medicamento':'olmesartan', 'dosis': 40, 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'}],
        'omesartan (olmesartán medoxomil) 20mg' : [{'medicamento':'olmesartan', 'dosis': 20, 'unidades': 'mg'}],
        'prazosin (prazosin) 1mg' : [{'medicamento':'prazosin', 'dosis': 1, 'unidades': 'mg'}],
        'prazosin (prazosin) 2mg' : [{'medicamento':'prazosin', 'dosis': 2, 'unidades': 'mg'}],
        'prazosina (prazosina) 1' : [{'medicamento':'prazosin', 'dosis': 1, 'unidades': 'mg'}],
        'prezoc (metoprolol succinato) 200 mg' : [{'medicamento':'metoprolol', 'dosis': 200, 'unidades': 'mg'}],
        'prezoc (metoprolol succinato) 25 mg' : [{'medicamento':'metoprolol', 'dosis': 25, 'unidades': 'mg'}],
        'prezoc (metoprolol succinato) 50 mg' : [{'medicamento':'metoprolol', 'dosis': 50, 'unidades': 'mg'}],
        'propanolol (propanolol clorhidrato) 40 mg' : [{'medicamento':'propanolol', 'dosis': 40, 'unidades': 'mg'}],
        'propanolol (propranolol clorhidrato) 80 mg' : [{'medicamento':'propanolol', 'dosis': 80, 'unidades': 'mg'}],
        'propranolol (propranolol clorhidrato) 40 mg' : [{'medicamento':'propanolol', 'dosis': 40, 'unidades': 'mg'}],
        'propranolol (propranolol clorhidrato) 80 mg' : [{'medicamento':'propanolol', 'dosis': 80, 'unidades': 'mg'}],
        'ramipril  (ramipril ) 5mg' : [{'medicamento':'ramipril', 'dosis': 5, 'unidades': 'mg'}],
        'satoren (losartan) 50mg' : [{'medicamento':'losartan', 'dosis': 50, 'unidades': 'mg'}],
        'tarka (trandolapril) 180mg' : [{'medicamento':'verapamilo', 'dosis': 180, 'unidades': 'mg'}, {'medicamento':'trandolapril', 'dosis': 2, 'unidades': 'mg'}],
        'tarka (verapamilo/trandolapril) 180mg/2mg' : [{'medicamento':'verapamilo', 'dosis': 180, 'unidades': 'mg'}, {'medicamento':'trandolapril', 'dosis': 2, 'unidades': 'mg'}],
        'telmisartan (telmisartan) 80mg' : [{'medicamento':'telmisartan', 'dosis': 80, 'unidades': 'mg'}],
        'telmisartan/amlodipino (telmisartan/amlodipino) 80/10' : [{'medicamento':'telmisartan', 'dosis': 80, 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}],
        'tenoretic (atenolol/clortalidona) 100/25 mg' : [{'medicamento':'atenolol', 'dosis': 100, 'unidades': 'mg'}, {'medicamento':'clortalidona', 'dosis': 25, 'unidades': 'mg'}],
        'tenormin (atenolol) 100 mg' : [{'medicamento':'atenolol', 'dosis': 100, 'unidades': 'mg'}],
        'tenormin (atenolol) 50 mg' : [{'medicamento':'atenolol', 'dosis': 50, 'unidades': 'mg'}],
        'teveten (eprosartán) 600 mg' : [{'medicamento':'eprosartan', 'dosis': 600, 'unidades': 'mg'}],
        'unipril (enalapril) 20mg'  : [{'medicamento':'enalapril', 'dosis': 20, 'unidades': 'mg'}],
        'valsarta/amlodipino/hidroclorotiazida () 320/10/25 mg' : [{'medicamento':'valsartan', 'dosis': 320, 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 25, 'unidades': 'mg'}],
        'valsartan  (valsartan ) 160mg' : [{'medicamento':'valsartan', 'dosis': 160, 'unidades': 'mg'}],
        'valsartan  (valsartan ) 80mg' : [{'medicamento':'valsartan', 'dosis': 80, 'unidades': 'mg'}],
        'valsartan (valsartan 320) 320mg' : [{'medicamento':'valsartan', 'dosis': 320, 'unidades': 'mg'}],
        'valsartan (valsartan) 80mg' : [{'medicamento':'valsartan', 'dosis': 80, 'unidades': 'mg'}],
        'valsartan+amlodipino (320/10) 320/10 mg' : [{'medicamento':'valsartan', 'dosis': 320, 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}],
        'valsartan/amlodipino (valsartan/amlodipino) 160/10' : [{'medicamento':'valsartan', 'dosis': 160, 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}],
        'valsartan/amlodipino (valsartan/amlodipino) 160/5 mg' : [{'medicamento':'valsartan', 'dosis': 160, 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'}],
        'valsartan/amlodipino (valsartan/amlodipino) 320/10' : [{'medicamento':'valsartan', 'dosis': 320, 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}],
        'valsartan/amlodipino/hidroclorotiazida () 320/10/25 mg' : [{'medicamento':'valsartan', 'dosis': 320, 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 25, 'unidades': 'mg'}],
        'valsartan/hidroclorotiazida/amlodipino (valsartan/hidroclorotiazida/amlodipino) 160/12.5/5' : [{'medicamento':'valsartan', 'dosis': 160, 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 12.5, 'unidades': 'mg'}],
        'varteral 160/10<br> (valsartán/amlodipino) 160/10' : [{'medicamento':'valsartan', 'dosis': 160, 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 10, 'unidades': 'mg'}],
        'varteral 320/5<br> (valsartán/amlodipino) 320/5' : [{'medicamento':'valsartan', 'dosis': 320, 'unidades': 'mg'}, {'medicamento':'amlodipino', 'dosis': 5, 'unidades': 'mg'}],
        'vasodyl (carvedilol) 25 mg' : [{'medicamento':'carvedilol', 'dosis': 25, 'unidades': 'mg'}],
        'verapamilo (verapamilo) 120mg ' : [{'medicamento':'verapamilo', 'dosis': 120, 'unidades': 'mg'}],
        'verapamilo (verapamilo) 40mg' : [{'medicamento':'verapamilo', 'dosis': 40, 'unidades': 'mg'}],
        'verapamilo (verapamilo) 40mg ' : [{'medicamento':'verapamilo', 'dosis': 40, 'unidades': 'mg'}],
        'verapamilo (verapamilo) 80mg ' : [{'medicamento':'verapamilo', 'dosis': 80, 'unidades': 'mg'}],
        'veratad   (verapamilo) 240mg' : [{'medicamento':'verapamilo', 'dosis': 240, 'unidades': 'mg'}],
        'ziac (bisoprolol fumarato/hidroclorotiazida) 10/6.25 mg' : [{'medicamento':'bisoprolol', 'dosis': 10, 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 6.25, 'unidades': 'mg'}],
        'ziac (bisoprolol fumarato/hidroclorotiazida) 2.5/6.25 mg' : [{'medicamento':'bisoprolol', 'dosis': 2.5, 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 6.25, 'unidades': 'mg'}],
        'ziac (bisoprolol fumarato/hidroclorotiazida) 5/6.25 mg' : [{'medicamento':'bisoprolol', 'dosis': 5, 'unidades': 'mg'}, {'medicamento':'hidroclorotiazida', 'dosis': 6.25, 'unidades': 'mg'}]
    }

    output = {}

    # en tratamiento es cuando al menos tiene un medicamento
    def entratamiento(lista_meds):
        return len(lista_meds) > 0


    lista_medicamentos_hipertension = []

    for medicamento in medicamentos:
        try:
            medicamento_hta = nombres_medicamentos[medicamento['nombre'].lower()]
            for med in medicamento_hta:
                lista_medicamentos_hipertension.append(med)
        except KeyError:
            continue

    return entratamiento(lista_medicamentos_hipertension)

## Generar el primer dataset para los pacientes
Este dataset incluye:
- numero identificador del paciente
- Fecha de nacimiento
- fecha de inicio de la prueba
- fecha de inicio de la noche (hora a la que el paciente se va a dormir)
- fecha de fin de la noche (hora a la que el paciente se despierta)
- genero (m = masculino, f = femenino)
- Talla (en metros)
- Peso (en Kg)
- Fuma (0 = no, 1 = si)
- frecuencia de ejercicio (1 = sedentario, 2 = menos de 70 minutos a la semana de actividad leve, 3 = entre 70 y 210 minutos a la semana, 4 = entre 210 y 500 minutos a la semana, 5 = mas de 500 minutos a la semana)
- consumo de cerveza (0 = no, 1 = si)
- consumo de vino (0 = no, 1 = si)
- consumo de licor fuerte (0 = no, 1 = si)
- consumo de cualquier tipo de licor (0 = no, 1 = si)
- medicamentos (lista que contiene los medicamentos que consume el paciente)

import pandas

data = []

for patient in patient_data:    
    patient_id = patient["id"]

    birth_date = patient["meta_data"][0]["fecha_nacimiento"]
    start_date = patient["meta_data"][0]["fecha_inicio"]

    start_night = patient["meta_data"][0]["inicio_noche"]
    end_night = patient["meta_data"][0]["fin_noche"]

    gender = patient["meta_data"][0]["genero"]
    height = patient["meta_data"][0]["talla"]
    weight = patient["meta_data"][0]["peso"]
    
    smoking = patient["meta_data"][0]["fuma"]
    excercise = patient["meta_data"][0]["frecuencia"]
    beer = patient["meta_data"][0]["cerveza"]
    wine = patient["meta_data"][0]["vino"]
    alcohol = patient["meta_data"][0]["licor"]
    strong_licor = patient["meta_data"][0]["licorFuerte"]
    treatment = hypertension_medicine(patient["drugs"])

    data.append([patient_id, treatment, birth_date, start_date, start_night, end_night, gender, height, weight, smoking, excercise, beer, wine, alcohol, strong_licor])
    

meta_data_df = pandas.DataFrame(
    data, 
    columns=[
        "patient_id",
        "treatment",
        "birth_date", 
        "start_date", 
        "start_night", 
        "end_night", 
        "gender", 
        "height", 
        "weight",
        "smoking",
        "excercise_frecuency",
        "beer",
        "wine",
        "alcohol",
        "strong_licor"
        
    ]
)

meta_data_df.set_index("patient_id", inplace=True)
meta_data_df.head()

## Generar el segundo dataset para los pacientes
Este dataset incluye:
- numero identificador del paciente
- fecha de la medicion
- presion sistolica (en mmHg)
- presion diastolica (en mmHg)
- frecuencia cardiaca (en latidos/min)

data = []

for patient in patient_data:   
    patient_id = patient["id"]
    
    for measure in patient["data"]:
        measure_date_time = measure["fecha_dt"]
        sistolic = measure["sistolica"]
        diastolic = measure["diastolica"]
        heart_reate = measure["valor"]
        
        data.append([patient_id, measure_date_time, sistolic, diastolic, heart_reate])
        
        
measure_df = pandas.DataFrame(
    data, 
    columns=[
        "patient_id",
        "measure_date_time", 
        "sistolic", 
        "diastolic", 
        "heart_rate",
    ]
)

measure_df.set_index("patient_id", inplace=True)
measure_df.head()

## Generar el tercer dataset para los pacientes
Este dataset incluye:
- numero identificador del paciente
- resistencia vascular sistemica
- presion de pulso
- velocidad de onda de pulso
- gasto cardiaco
- indice cardiaco
- volumen latido
- presion arterial media

data = []

for patient in patient_data: 
    intermediary = []
    
    intermediary.append(patient["id"])
    for item in ['RVS', 'PP', 'VOP', 'GC', 'IC', 'VL', 'PAM']:
        intermediary.append(patient["measure"]["diagnostico"]["output"][item])
    data.append(intermediary)
    

hemodinamic_data_df = pandas.DataFrame(
    data, 
    columns=[
        'patient_id',
        'systemic_vascular_resistance',
        'pulse_pressure',
        'pulse_wave_velocity',     
        'cardiac_output',
        'cardiac_index',
        'stroke_volume',
        'mean_arterial_pressure'
    ]
)

hemodinamic_data_df.set_index("patient_id", inplace=True)
hemodinamic_data_df.head()

## Concatenar los datasets

patient_df = pandas.merge(
    meta_data_df, 
    measure_df, 
    how="inner", 
    left_index=True, 
    right_index=True
)

patient_df.head()

final_df = pandas.merge(
    patient_df, 
    hemodinamic_data_df, 
    how="inner", 
    left_index=True, 
    right_index=True
)

final_df.head()

final_df.to_csv('base_dataset.csv')
