==> 6.0 Descomposición en Valores Singulares.ipynb <==
# Descomposición en Valores Singulares (SVD)

## Introducción

Cuando trabajamos en problemas de Machine Learning, muchas veces nos encontramos con conjuntos de datos grandes que contienen además muchas características o feactures. Una forma simple de reducir las dimensiones de estas características es aplicar alguna técnica de Factorización de matrices.

La factorización es una técnica que consiste en la descomposición de una expresión matemática (que puede ser un número, una matriz, un tensor, etc.) en forma de producto. La descomposición en valores singulares, es la base del análisis de componentes principales (PCA), cuyo propósito es derivar un número relativamente pequeño de combinaciones lineales no correlacionadas (componentes principales) de un conjunto de variables aleatorias de media cero mientras que conserva la mayor cantidad de información de las variables originales como sea posible. 

Los principales objetivos del PCA son:

- Reducción de dimensionalidad
- Determinación de combinaciones lineales de variables
- Selección de características o features más útiles
- Visualización de datos multidimensionales
- Identificación de las variables subyacentes
- Identificación grupos de objetos o de valores atípicos

La descomposición en valores singulares para una matriz de datos A está dada por la expresión:

$$A = U \Sigma V^T$$

donde: 

- U es una matriz mxm cuyas columnas corresponden a los vectores singulares izquierdos de A
- $\Sigma$ es una matriz mxn diagonal cuyas entradas corresponden a los valores singulares de A.
- $V^T$ es la traspuesta de la matriz nxn T cuyas columnas corresponden a los vectores singulares derechos de A.

%matplotlib inline

import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

sns.set()

#Lectura de datos
datos= pd.read_csv('data/filtered1_dataset.csv', sep=',', index_col="patient_id")
datos.head()

# Extraemos la variable respuesta "treatment"
y = pd.DataFrame(datos["treatment"])
y.head()

np.shape(y)

# Extraer las variables independientes 
X = datos.iloc[:,6:]
X.head()

np.shape(X)

# Singular-value decomposition, usando la función svd()
from scipy.linalg import svd
U, s, VT = svd(X)
print('Matriz U')
print(U)
print('Matriz Sigma')
print(s)
print('Matriz VT')
print(VT)

U.shape, VT.shape, s.shape

## Reconstruir matriz desde SVD

La matriz original se puede reconstruir a partir de los elementos $U$, $\Sigma$ y $V ^ T$. Pero los elementos $U$, $\Sigma$ y $V ^ T$ devueltos desde la función svd() no se pueden multiplicar directamente.

El vector s debe convertirse en una matriz diagonal utilizando la función diag(). Por defecto, esta función creará una matriz cuadrada que es n x n, en relación con nuestra matriz original. Esto causa un problema ya que el tamaño de las matrices no se ajusta a las reglas de multiplicación de matrices, donde el número de columnas en una matriz debe coincidir con el número de filas en la matriz posterior.

Después de crear la matriz diagonal $\Sigma$ cuadrada, los tamaños de las matrices son relativos a la matriz original m x n que estamos descomponiendo, de la siguiente manera:

# Reconstruir la matriz de datos
from numpy import array
from numpy import diag
from numpy import dot
from numpy import zeros
from scipy.linalg import svd

# Creamos la matriz Sigma mxn 
S = zeros((X.shape[0], X.shape[1]))
S[:X.shape[1], :X.shape[1]] = diag(s)
datos1 = U.dot(S.dot(VT))
print('Matriz de datos reconstruida')
pd.DataFrame(datos1.round(2), columns=X.columns, index=X.index)

Se observa que la matriz de datos se puede descomponer como el producto de las matrices  $U$, $\Sigma$ y $V^T$. Esto indica que cualquier transformación podría expresarse como una secuencia de rotación $V^T$, escala $\Sigma$ y una nueva rotación $U$, es decir, cada valor es un factor de escala a lo largo de un eje particular.

En el cado de la Clasificación y/o regresión estos valores indican el impacto de un eje particular en el resultado general.
